Quiz class
generate a quiz game object and give it attributes of question number (default of 0), score (default of 0), question list (as an input) and current question (default None).

The question methods: still_has_questions is used to monitor whether or not there are remaining questions to be asked,.

next_question goes through the question list, presents the user with a question and asks for a user response (T/F)

check_answer checks the user answer against the actual answer and updates their score accordingly.